+++ games/games.mjs
@@ -2,16 +2,17 @@
 export const GameConstants = [
   {
     title: "2022",
-    filename: "2022-field.jpg",
-    source: "https://firstfrc.blob.core.windows.net/frc2022/Manual/2022FRCGameManual.pdf",
-    topLeft: [149, 24],
-    bottomRight: [1043, 472],
+    imageLight: "2022-field-light.png",
+    imageDark: "2022-field-dark.png",
+    source: "https://www.chiefdelphi.com/t/2022-top-down-field-renders/399031",
+    topLeft: [441, 74],
+    bottomRight: [3082, 1394],
     widthInches: 648,
     heightInches: 324
   },
   {
     title: "2020",
-    filename: "2020-field.png",
+    imageLight: "2020-field.png",
     source: "https://www.chiefdelphi.com/t/2020-top-down-orthographic-field-views/369892",
     topLeft: [519, 73],
     bottomRight: [3028, 1363],
@@ -20,7 +21,7 @@ export const GameConstants = [
   },
   {
     title: "2019",
-    filename: "2019-field.jpg",
+    imageLight: "2019-field.jpg",
     source: "https://github.com/wpilibsuite/PathWeaver/blob/main/src/main/resources/edu/wpi/first/pathweaver/2019-field.jpg",
     topLeft: [212, 34],
     bottomRight: [1370, 607],
@@ -29,7 +30,7 @@ export const GameConstants = [
   },
   {
     title: "2018",
-    filename: "2018-field.jpg",
+    imageLight: "2018-field.jpg",
     source: "https://firstfrc.blob.core.windows.net/frc2018/Manual/2018FRCGameSeasonManual.pdf",
     topLeft: [248, 35],
     bottomRight: [1588, 710],
@@ -38,7 +39,7 @@ export const GameConstants = [
   },
   {
     title: "2017",
-    filename: "2017-field.jpg",
+    imageLight: "2017-field.jpg",
     source: "https://firstfrc.blob.core.windows.net/frc2017/Manual/2017FRCGameSeasonManual.pdf",
     topLeft: [198, 39],
     bottomRight: [1090, 480],
@@ -47,7 +48,7 @@ export const GameConstants = [
   },
   {
     title: "2016",
-    filename: "2016-field.jpg",
+    imageLight: "2016-field.jpg",
     source: "https://firstfrc.blob.core.windows.net/frc2016manuals/GameManual/FRC-2016-game-manual.pdf",
     topLeft: [124, 48],
     bottomRight: [791, 375],