+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-Added a 2022 field image extracted from the manual. This is likely to change if better renders are available.
+* Added support for light and dark field images.
+* Updated the 2022 field images to [these renders](https://www.chiefdelphi.com/t/2022-top-down-field-renders/399031).